# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: HackJudge API
  description: Production-ready hackathon judging platform API
  version: 1.0.0
servers:
  - url: http://localhost:3001/api/v1
paths:
  /health:
    get:
      summary: Health check
      responses:
        "200":
          description: OK
  /auth/magic-link:
    post:
      summary: Send magic link to judge
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                eventSlug: { type: string }
      responses:
        "200":
          description: Magic link sent
        "429":
          description: Rate limited
  /auth/verify/{token}:
    get:
      summary: Verify magic link
      parameters:
        - name: token
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: JWT pair returned
        "401":
          description: Invalid, used, or expired token
  /auth/refresh:
    post:
      summary: Refresh access token
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken: { type: string }
      responses:
        "200":
          description: New access token
  /auth/organizer/login:
    post:
      summary: Organizer login
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        "200":
          description: JWT pair returned
  /events:
    post:
      summary: Create event from YAML
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                configYaml: { type: string }
      responses:
        "201":
          description: Event created
        "400":
          description: Validation error
        "409":
          description: Slug collision
  /events/{slug}:
    get:
      summary: Get event details
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Event details
        "404":
          description: Event not found
  /events/{slug}/status:
    get:
      summary: Get event status and completion stats
      parameters:
        - name: slug
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Status and stats
  /events/{slug}/teams:
    get:
      summary: List teams
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Teams list
  /events/{slug}/scores:
    put:
      summary: Upsert scores
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                teamId: { type: string }
                scores:
                  type: array
                  items:
                    type: object
                    properties:
                      criterionId: { type: string }
                      value: { type: number }
      responses:
        "200":
          description: Scores saved
        "403":
          description: Judging closed or locked
    get:
      summary: Get all scores (organizer)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: All scores
  /events/{slug}/scores/submit:
    post:
      summary: Mark team as fully scored
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                teamId: { type: string }
      responses:
        "200":
          description: Submitted
  /events/{slug}/results/generate:
    post:
      summary: Generate results snapshot
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Results generated
  /events/{slug}/results:
    get:
      summary: Get latest results
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Results snapshot
  /events/{slug}/results/export/csv:
    get:
      summary: Export results as CSV
      security:
        - bearerAuth: []
      responses:
        "200":
          description: CSV stream
          content:
            text/csv:
              schema: { type: string }
  /events/{slug}/results/export/pdf:
    get:
      summary: Export certificates as PDF
      security:
        - bearerAuth: []
      responses:
        "200":
          description: PDF stream
          content:
            application/pdf:
              schema: { type: string }
  /events/{slug}/results/export/xlsx:
    get:
      summary: Export results as Excel (.xlsx)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Excel stream
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema: { type: string }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT